requirements.txt
amqp>=5.2.0,<6.0
appdirs==1.4.4
APScheduler==3.10.4
billiard>=4.2.0,<5.0
celery>=5.4.0,<6.0
coreapi>=2.3.3
Django==4.2.16
django-appconf==1.0.6
django-codemirror2==0.2
django-colorfield==0.1.15
django-cors-headers==3.14.0
django-filter==2.4.0
django-guardian==2.4.0
django-imagekit==4.1.0
django-redis==5.4.0
django-webpack-loader==3.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
djangorestframework-guardian==0.3.0
drf-nested-routers==0.95.0
funcsigs==1.0.2
futures==3.1.1
gunicorn==23.0.0
geodeep==0.9.8
itypes==1.1.0
kombu>=5.3.0,<6.0
Markdown==3.5.2
olefile==0.44
openapi-codec==1.1.7
packaging==24.1
piexif==1.0.13
pilkit==2.0
Pillow==8.4.0
pip-autoremove==0.9.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
pydantic==1.10.8
pyodm==1.5.11
pyparsing==3.1.2
pytz==2024.1
rcssmin==1.0.6
redis==5.0.4
requests-toolbelt==1.0.0
requests>=2.21.0
rfc3987==1.3.7
rjsmin==1.0.12
simplejson==3.10.0
six==1.16.0
strict-rfc3339==0.7
tzlocal==2.1
uritemplate==3.0.0
vine>=5.1.0,<6.0
webcolors==1.5
rio-tiler==2.1.2
rio-color==1.0.4
rio-cogeo==2.3.1
rasterio==1.2.9 ; sys_platform == 'linux' or sys_platform == 'darwin'
https://github.com/OpenDroneMap/WebODM/releases/download/v1.9.7/rasterio-1.2.10-cp39-cp39-win_amd64.whl ; sys_platform == "win32"
https://github.com/OpenDroneMap/WebODM/releases/download/v1.9.7/GDAL-3.3.3-cp39-cp39-win_amd64.whl ; sys_platform == "win32"
Shapely==1.8.0 ; sys_platform == "win32"
eventlet==0.32.0 ; sys_platform == "win32"
pyopenssl==24.0.0 ; sys_platform == "win32"
numpy==1.26.4
scipy==1.11.4
drf-yasg==1.20.0
zipstream-ng==1.8.0
